rename job package to batches
