+++ api/batches/batch.go
@@ -1,4 +1,4 @@
-package jobs
+package batches
 
 import (
 	"bytes"
