+++ api/batches/batch_ent.go
@@ -1,4 +1,4 @@
-package jobs
+package batches
 
 import (
 	"strings"
